==> notebooks/03_match_cts_extent.Rmd <==

```{r}
library(tidyverse)
library(magrittr)
library(lubridate)
library(sf)
library(raster)
#library(rgeos)
library(viridis)
library(tictoc)
```

## Match the crs and extents of sf objects

```{r}
# Load 10 km grid
grid_sf = read_sf("../data/input/local_data/10km_grid_wgs84/10km_grid_wgs84.shp")
zip_sf = read_sf("../data/input/local_data/Zipcode_Info/polygon/ESRI06USZIP5_POLY_WGS84.shp")

st_crs(grid_sf)
st_crs(zip_sf)
```

```{r}
grid_sf %>% 
  st_simplify() %>% 
  ggplot(aes(fill = "red"), alpha = 0.75, lwd = 0.1) + 
  geom_sf() + 
  theme(legend.position = "none")
```

```{r}
zip_sf %>% 
  ggplot() + 
  geom_sf(aes(fill = "red"), alpha = 0.75, lwd = 0.1) + 
  theme(legend.position = "none")
```

```{r}
ext.ras <- extent(grid_sf)
ext.pol <- extent(zip_sf)

plot(ext.ras, 
     xlim = c(min(ext.ras@xmin, ext.pol@xmin), 
              max(ext.ras@xmax, ext.pol@xmax)), 
     ylim= c(min(ext.ras@ymin, ext.pol@ymin), 
              max(ext.ras@ymax, ext.pol@ymax)), 
     col="red")
plot(ext.pol, add=T, col="blue")
```

```{r}
sf_use_s2(FALSE)
zip_sf <- st_crop(st_simplify(zip_sf), st_bbox(grid_sf))
```

```{r}
zip_sf %>% 
  ggplot() + 
  geom_sf(aes(fill = "red"), alpha = 0.75, lwd = 0.1) + 
  theme(legend.position = "none")
```

```{r}
ext.ras <- extent(grid_sf)
ext.pol <- extent(zip_sf)

plot(ext.ras, 
     xlim = c(min(ext.ras@xmin, ext.pol@xmin), 
              max(ext.ras@xmax, ext.pol@xmax)), 
     ylim= c(min(ext.ras@ymin, ext.pol@ymin), 
              max(ext.ras@ymax, ext.pol@ymax)), 
     col="red")
plot(ext.pol, add=T, col="blue")
```

```{r}
## save matching sf's
# write_rds(grid_sf, "../data/intermediate/scratch/grid_sf.rds")
# write_rds(zip_sf, "../data/intermediate/scratch/zip_sf.rds")
```
