==> notebooks/04_test_grid_to_zip.Rmd <==

```{r, warning=FALSE, message=FALSE}
library(tidyverse)
library(magrittr)
library(lubridate)
library(sf)
library(raster)
library(rgeos)
library(viridis)
library(tictoc)
```

## Join smoke values to grid

```{r}
## Load grid and zip sf objects
grid_sf = read_rds("../data/intermediate/scratch/grid_sf.rds") %>% 
  rename(grid_id = ID)
zip_sf = read_rds("../data/intermediate/scratch/zip_sf.rds") 

# Load smoke grid data frame
smoke_grid_df_list = read_rds("../data/intermediate/scratch/smoke_grid_df_list.rds")

smoke_grid_df = smoke_grid_df_list[["2006"]]
head(smoke_grid_df)
```

```{r}
# Join grid with smoke values
grid_sf %<>% 
  left_join(smoke_grid_df)

sum(!grid_sf$grid_id %in% smoke_grid_df$grid_id)
sum(is.na(grid_sf$smoke))
```

```{r}
grid_sf %>% 
  #st_as_sf() %>% 
  st_simplify() %>% 
  ggplot() + 
  geom_sf(aes(fill = smoke), colour = NA, lwd = 0) + 
  scale_fill_viridis()
```

## Test `over`

```{r}
zip_i_sf <- dplyr::select(zip_sf[1, ], c("geometry"))
grid_i_sf <- dplyr::select(st_crop(grid_sf, extent(zip_i_sf)), c("smoke", "geometry"))

grid_i_sf %>% 
  ggplot() + 
  geom_sf(aes(fill = smoke), colour = NA, lwd = 0) + 
  scale_fill_viridis() + 
  geom_sf(data = zip_i_sf, linewidth = 1, color = "red", fill = NA)
```

```{r}
zip_i_sp = as_Spatial(zip_i_sf)
grid_i_sp = as_Spatial(grid_i_sf)

class(zip_i_sp)
class(grid_i_sp)

as.numeric(zip_i_sp %over% grid_i_sp)
```

## Measure time it takes to apply `over` to 100 zips

```{r, warning = F, message= F}
zip_sf$smoke = 0

## iterating over zip codes ----
tic("apply *over* to 100 zips")
for (i in 1:100) {
  zip_i_sf <- dplyr::select(zip_sf[i, ], c("geometry"))
  grid_i_sf <- try(
    dplyr::select(st_crop(grid_sf, extent(zip_i_sf)), c("smoke", "geometry")),
    silent = T)

  if(class(grid_i_sf)[1] != "try-error"){
    zip_i_sp = as_Spatial(zip_i_sf)
    grid_i_sp = as_Spatial(grid_i_sf)

    zip_sf$smoke[i] <- as.numeric(zip_i_sp %over% grid_i_sp)
  } else zip_sf$smoke[i] <- as.numeric(NA)
}
toc()
```

```{r}
## plot 100 zips ----
zip_sf %>% 
  filter(smoke > 0) %>%  
  st_simplify() %>% 
  ggplot() + 
  geom_sf(data = zip_sf, lwd = 0.1) + 
  geom_sf(aes(fill = smoke), lwd = 0.1) + 
  scale_fill_viridis()
```

## Test `var_poly_to_poly` function on 10 zips

```{r, warning = F, message= F}
source("../../lib/var_poly_to_poly.R")

var_poly_to_poly(
  x_poly_sf = grid_sf, 
  y_poly_sf = zip_sf[1:10, ], 
  y_id = "ZIP", 
  x_var = "smoke", 
  fn = mean
)
```
